changes to migration script so seen defaults to false removed duplication in DatabaseViewsExt extracted PKReminder class to its own file fixed some privacy stuff so you can only ask for your own system's reminders
