+++ PluralKit.Bot/Commands/Lists/ContextListExt.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -167,10 +167,10 @@ namespace PluralKit.Bot
             }
         }
         public static async Task RenderReminderList(this Context ctx, IDatabase db, MemberId member, string embedTitle, string color, bool showSeen) {
-            var reminders = await db.Execute(conn => showSeen ? conn.QueryReminders(member) : conn.QueryUnseenReminders(member)).ToListAsync();
+            var reminders = await db.Execute(conn => showSeen ? conn.QueryReminders(member) : conn.QueryReminders(member, false)).ToListAsync();
             var itemsPerPage = 25;
 
-            if (reminders.Count > 0 && showSeen) {
+            if (reminders.Count > 0 || showSeen) {
                 await ctx.Paginate(reminders.ToAsyncEnumerable(), reminders.Count, itemsPerPage, embedTitle, color, (eb, page) => {
                     eb.Footer(new("result".ToQuantity(reminders.Count)));
                     eb.WithSimpleLineContent(page.Select(r => {
