+++ PluralKit.Bot/Commands/Member.cs
@@ -76,7 +76,7 @@ namespace PluralKit.Bot
         public async Task AddReminder(Context ctx, PKMember target) {
             if (ctx.System?.Id == target.System) {
                 await using var conn = await _db.Obtain();
-                await _repo.AddReminder(conn, new PKReminder { Mid = ctx.Message.Id, Channel = ctx.Channel.Id, Guild = ctx.Guild == null ? null : ctx.Guild.Id, Member = target.Id, System = target.System, Seen = false });
+                await _repo.AddReminder(conn, new PKReminder { Mid = ctx.Message.Id, Channel = ctx.Channel.Id, Guild = ctx.Guild == null ? null : ctx.Guild.Id, Member = target.Id, System = target.System });
                 await ctx.Reply($"Added new reminder for {target.Name}");
             }
             else {
@@ -84,12 +84,20 @@ namespace PluralKit.Bot
             }
         }
 
-        public async Task GetReminders(Context ctx, PKMember target) => await ctx.RenderReminderList(
-            _db,
-            target.Id,
-            $"Reminders for {target.Name}",
-            target.Color,
-            true);
+        public async Task GetReminders(Context ctx, PKMember target) {
+            if (ctx.System?.Id == target.System) {
+                await ctx.RenderReminderList(
+                    _db,
+                    target.Id,
+                    $"Reminders for {target.Name}",
+                    target.Color,
+                    true);
+            }
+            else {
+                await ctx.Reply($"{Emojis.Error} You can only view reminders from your own system.");
+            }
+        }
+
         public async Task Soulscream(Context ctx, PKMember target)
         {
             // this is for a meme, please don't take this code seriously. :)
