+++ PluralKit.Core/Database/Views/DatabaseViewsExt.cs
@@ -46,25 +46,19 @@ namespace PluralKit.Core
         }
 
         //r1 and r2 are needed in order to return the "seen" value before it's updated
-        public static IAsyncEnumerable<PKReminder> QueryUnseenReminders(this IPKConnection conn, MemberId member) => conn.QueryStreamAsync<PKReminder>(@"
+        public static IAsyncEnumerable<PKReminder> QueryReminders(this IPKConnection conn, MemberId member, bool seen = true) {
+            var showSeen = seen ? "" : "AND seen = false";
+            var query = @$"
 WITH x AS (
     UPDATE reminders r1
     SET seen = true
-    FROM (SELECT mid, channel, guild, member, system, seen, timestamp FROM reminders WHERE member = @Id AND seen = false) r2
+    FROM (SELECT mid, channel, guild, member, system, seen, timestamp FROM reminders WHERE member = @Id {showSeen}) r2
     WHERE r1.mid = r2.mid
     RETURNING r2.*
 )
-SELECT * FROM x ORDER BY timestamp DESC", new { Id = member.Value });
-
-        public static IAsyncEnumerable<PKReminder> QueryReminders(this IPKConnection conn, MemberId member) => conn.QueryStreamAsync<PKReminder>(@"
-WITH x AS (
-    UPDATE reminders r1
-    SET seen = true
-    FROM (SELECT mid, channel, guild, member, system, seen, timestamp FROM reminders WHERE member = @Id) r2
-    WHERE r1.mid = r2.mid
-    RETURNING r2.*
-)
-SELECT * FROM x ORDER BY timestamp DESC", new { Id = member.Value });
+SELECT * FROM x ORDER BY timestamp DESC";
+            return conn.QueryStreamAsync<PKReminder>(query, new { Id = member.Value });
+        }
 
 
         public struct MemberListQueryOptions
